% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_characteristics}
\alias{list_characteristics}
\alias{object_list}
\title{Show lists of models and characteristics}
\usage{
list_characteristics()

object_list()
}
\value{
\code{list_characteristics()} returns a tibble with columns
\code{characteristic} and \code{object}. \code{object_list()} returns a character string.
}
\description{
Show lists of models and characteristics
}
