% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_active.R
\name{pluck_features_active}
\alias{pluck_features_active}
\alias{.pluck_features_active}
\alias{.pluck_features_active.default}
\alias{.pluck_features_active.workflow}
\alias{.pluck_features_active.model_fit}
\alias{.pluck_features_active.terms}
\alias{.pluck_features_active.rpart}
\alias{.pluck_features_active.ranger}
\alias{.pluck_features_active.xgb.Booster}
\alias{.pluck_features_active.glmnet}
\alias{.pluck_features_active.tidy_cubist}
\alias{.pluck_features_active.cubist}
\alias{.pluck_features_active.tidy_C50}
\alias{.pluck_features_active.C5.0}
\alias{.pluck_features_active.earth}
\alias{.pluck_features_active.mixo_spls}
\alias{.pluck_features_active.mixo_pls}
\alias{.pluck_features_active.mixo_splsda}
\alias{.pluck_features_active.mixo_plsda}
\alias{.pluck_features_active.bagger}
\alias{.pluck_features_active.bart}
\alias{.pluck_features_active.party}
\alias{.pluck_features_active.cforest}
\alias{.pluck_features_active.tidy_xrf}
\alias{.pluck_features_active.xrf}
\alias{.pluck_features_active.lgb_trees}
\alias{.pluck_features_active.lgb.Booster}
\alias{.pluck_features_active.nullmodel}
\alias{.pluck_features_active.lm}
\alias{.pluck_features_active.glm}
\alias{.pluck_features_active.kknn}
\alias{.pluck_features_active.stanreg}
\alias{.pluck_features_active.hurdle}
\alias{.pluck_features_active.zeroinfl}
\alias{.pluck_features_active.gam}
\alias{.pluck_features_active.lda}
\alias{.pluck_features_active.fda}
\alias{.pluck_features_active.gen.ridge}
\alias{.pluck_features_active.qda}
\alias{.pluck_features_active.rda}
\alias{.pluck_features_active.randomForest}
\alias{.pluck_features_active.lda_diag}
\alias{.pluck_features_active.qda_diag}
\alias{.pluck_features_active._keras.engine.sequential.Sequential}
\alias{.pluck_features_active.LiblineaR}
\alias{.pluck_features_active.mda}
\alias{.pluck_features_active.naive_bayes}
\alias{.pluck_features_active.NaiveBayes}
\alias{.pluck_features_active.brulee_multinomial_reg}
\alias{.pluck_features_active.brulee_logistic_reg}
\alias{.pluck_features_active.brulee_linear_reg}
\alias{.pluck_features_active.brulee_mlp}
\alias{.pluck_features_active.sda}
\title{Predictors used in the model}
\usage{
.pluck_features_active(x, ...)

\method{.pluck_features_active}{default}(x, ...)

\method{.pluck_features_active}{workflow}(x, ...)

\method{.pluck_features_active}{model_fit}(x, ...)

\method{.pluck_features_active}{terms}(x, ...)

\method{.pluck_features_active}{rpart}(x, ...)

\method{.pluck_features_active}{ranger}(x, ...)

\method{.pluck_features_active}{xgb.Booster}(x, nrounds = x$niter, ...)

\method{.pluck_features_active}{glmnet}(x, penalty = 0.001, ...)

\method{.pluck_features_active}{tidy_cubist}(x, committees = max(x$committee), ...)

\method{.pluck_features_active}{cubist}(x, committees = x$committees, ...)

\method{.pluck_features_active}{tidy_C50}(x, trials = max(x$trial), ...)

\method{.pluck_features_active}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{.pluck_features_active}{earth}(x, ...)

\method{.pluck_features_active}{mixo_spls}(x, ...)

\method{.pluck_features_active}{mixo_pls}(x, ...)

\method{.pluck_features_active}{mixo_splsda}(x, ...)

\method{.pluck_features_active}{mixo_plsda}(x, ...)

\method{.pluck_features_active}{bagger}(x, ...)

\method{.pluck_features_active}{bart}(x, ...)

\method{.pluck_features_active}{party}(x, ...)

\method{.pluck_features_active}{cforest}(x, ...)

\method{.pluck_features_active}{tidy_xrf}(x, ...)

\method{.pluck_features_active}{xrf}(x, penalty = 0.001, ...)

\method{.pluck_features_active}{bagger}(x, ...)

\method{.pluck_features_active}{lgb_trees}(x, trees = max(x$trees), ...)

\method{.pluck_features_active}{lgb.Booster}(x, trees = x$params$num_iterations, ...)

\method{.pluck_features_active}{nullmodel}(x, ...)

\method{.pluck_features_active}{lm}(x, ...)

\method{.pluck_features_active}{glm}(x, ...)

\method{.pluck_features_active}{kknn}(x, ...)

\method{.pluck_features_active}{stanreg}(x, ...)

\method{.pluck_features_active}{hurdle}(x, ...)

\method{.pluck_features_active}{zeroinfl}(x, ...)

\method{.pluck_features_active}{gam}(x, ...)

\method{.pluck_features_active}{lda}(x, ...)

\method{.pluck_features_active}{fda}(x, ...)

\method{.pluck_features_active}{gen.ridge}(x, ...)

\method{.pluck_features_active}{qda}(x, ...)

\method{.pluck_features_active}{rda}(x, ...)

\method{.pluck_features_active}{randomForest}(x, ...)

\method{.pluck_features_active}{lda_diag}(x, ...)

\method{.pluck_features_active}{qda_diag}(x, ...)

\method{.pluck_features_active}{`_keras.engine.sequential.Sequential`}(x, ...)

\method{.pluck_features_active}{LiblineaR}(x, ...)

\method{.pluck_features_active}{mda}(x, ...)

\method{.pluck_features_active}{naive_bayes}(x, ...)

\method{.pluck_features_active}{NaiveBayes}(x, ...)

\method{.pluck_features_active}{brulee_multinomial_reg}(x, ...)

\method{.pluck_features_active}{brulee_logistic_reg}(x, ...)

\method{.pluck_features_active}{brulee_linear_reg}(x, ...)

\method{.pluck_features_active}{brulee_mlp}(x, ...)

\method{.pluck_features_active}{sda}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{penalty}{A numeric parameter value for the amount of
regularization/penalty of the model.}

\item{committees}{The number of Cubist committees to use in the characterization.}

\item{trials, nrounds, trees}{The number of boosting to use in the characterization.}
}
\description{
\code{.pluck_features_active()} enumerates all of the features/predictors that
have contributions to the prediction equation. This excludes variables that
were included in the model fit but have not affect on the prediction
equations. For example, if a decision tree never split on a particular
feature, it is not included in this list.
}
\details{
Note that the term "features" corresponds to predictors produced by any
pre-processing operations executed by the model (if any). For example, if
\code{lm(y + x + z, data)} is used and predictor \code{x} is a factor, some models
will convert \code{x} to a set of binary indicator columns. In that instance, this
function will list the indicator columns (instead of \code{x}).
}
\keyword{internal}
