% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_input.R
\name{pluck_features_input}
\alias{pluck_features_input}
\alias{.pluck_features_input}
\alias{.pluck_features_input.default}
\alias{.pluck_features_input.workflow}
\alias{.pluck_features_input.model_fit}
\alias{.pluck_features_input.recipe}
\alias{.pluck_features_input.terms}
\alias{.pluck_features_input.formula}
\alias{.pluck_features_input.keras.engine.sequential.Sequential}
\alias{.pluck_features_input._keras.engine.sequential.Sequential}
\alias{.pluck_features_input.glmnet}
\alias{.pluck_features_input.multnet}
\title{Predictors given in the model}
\usage{
.pluck_features_input(x, ...)

\method{.pluck_features_input}{default}(x, ...)

\method{.pluck_features_input}{workflow}(x, include_intercept = FALSE, ...)

\method{.pluck_features_input}{model_fit}(x, ...)

\method{.pluck_features_input}{recipe}(x, ...)

\method{.pluck_features_input}{terms}(x, ...)

\method{.pluck_features_input}{formula}(x, data, ...)

\method{.pluck_features_input}{keras.engine.sequential.Sequential}(x, ...)

\method{.pluck_features_input}{`_keras.engine.sequential.Sequential`}(x, ...)

\method{.pluck_features_input}{glmnet}(x, ...)

\method{.pluck_features_input}{multnet}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{data}{A data frame of the source data.}
}
\description{
\code{.pluck_features_input()} enumerates all of the features/predictors that
have or will be passed to a modeling function.
}
\keyword{internal}
