% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_rule_size.R
\name{pluck_mean_rule_size}
\alias{pluck_mean_rule_size}
\alias{.pluck_mean_rule_size}
\alias{.pluck_mean_rule_size.default}
\alias{.pluck_mean_rule_size.workflow}
\alias{.pluck_mean_rule_size.model_fit}
\alias{.pluck_mean_rule_size.tidy_C50}
\alias{.pluck_mean_rule_size.C5.0}
\alias{.pluck_mean_rule_size.tidy_cubist}
\alias{.pluck_mean_rule_size.cubist}
\alias{.pluck_mean_rule_size.tidy_xrf}
\alias{.pluck_mean_rule_size.xrf}
\title{Compute the mean number of predictor in rules in a rule-based model}
\usage{
.pluck_mean_rule_size(x, ...)

\method{.pluck_mean_rule_size}{default}(x, ...)

\method{.pluck_mean_rule_size}{workflow}(x, ...)

\method{.pluck_mean_rule_size}{model_fit}(x, ...)

\method{.pluck_mean_rule_size}{tidy_C50}(x, trials = max(x$trial), ...)

\method{.pluck_mean_rule_size}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{.pluck_mean_rule_size}{tidy_cubist}(x, committees = max(x$committee), ...)

\method{.pluck_mean_rule_size}{cubist}(x, committees = x$committees, ...)

\method{.pluck_mean_rule_size}{tidy_xrf}(x, ...)

\method{.pluck_mean_rule_size}{xrf}(x, penalty = 0.001, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{committees}{The number of Cubist committees to use in the examination.}

\item{penalty}{A numeric parameter value for the amount of
regularization/penalty of the model.}
}
\description{
Compute the mean number of predictor in rules in a rule-based model
}
\keyword{internal}
