% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_input.R, R/num_features_input.R
\name{.pluck_features_input.hardhat_blueprint}
\alias{.pluck_features_input.hardhat_blueprint}
\alias{pluck_num_features_input}
\alias{.pluck_num_features_input}
\alias{.pluck_num_features_input.default}
\alias{.pluck_num_features_input.workflow}
\alias{.pluck_num_features_input.model_fit}
\title{Number of predictors given to the model}
\usage{
\method{.pluck_features_input}{hardhat_blueprint}(x, include_intercept = FALSE, ...)

.pluck_num_features_input(x, ...)

\method{.pluck_num_features_input}{default}(x, ...)

\method{.pluck_num_features_input}{workflow}(x, ...)

\method{.pluck_num_features_input}{model_fit}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}
}
\description{
Number of predictors given to the model
}
\keyword{internal}
