% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_parameters.R
\name{pluck_num_parameters}
\alias{pluck_num_parameters}
\alias{.pluck_num_parameters}
\alias{.pluck_num_parameters.default}
\alias{.pluck_num_parameters.workflow}
\alias{.pluck_num_parameters.model_fit}
\alias{.pluck_num_parameters.tidy_cubist}
\alias{.pluck_num_parameters.cubist}
\alias{.pluck_num_parameters.multinom}
\alias{.pluck_num_parameters.nnet}
\alias{.pluck_num_parameters.glmnet}
\alias{.pluck_num_parameters.earth}
\alias{.pluck_num_parameters.mixo_spls}
\alias{.pluck_num_parameters.mixo_pls}
\alias{.pluck_num_parameters.mixo_splsda}
\alias{.pluck_num_parameters.mixo_plsda}
\alias{.pluck_num_parameters.tidy_xrf}
\alias{.pluck_num_parameters.xrf}
\alias{.pluck_num_parameters.bagger}
\title{Compute the number of estimated model parameters}
\usage{
.pluck_num_parameters(x, ...)

\method{.pluck_num_parameters}{default}(x, ...)

\method{.pluck_num_parameters}{workflow}(x, ...)

\method{.pluck_num_parameters}{model_fit}(x, ...)

\method{.pluck_num_parameters}{tidy_cubist}(x, committees = max(x$committee), ...)

\method{.pluck_num_parameters}{cubist}(x, committees = x$committees, ...)

\method{.pluck_num_parameters}{multinom}(x, ...)

\method{.pluck_num_parameters}{nnet}(x, ...)

\method{.pluck_num_parameters}{glmnet}(x, penalty = 0.001, ...)

\method{.pluck_num_parameters}{earth}(x, ...)

\method{.pluck_num_parameters}{mixo_spls}(x, ...)

\method{.pluck_num_parameters}{mixo_pls}(x, ...)

\method{.pluck_num_parameters}{mixo_splsda}(x, ...)

\method{.pluck_num_parameters}{mixo_plsda}(x, ...)

\method{.pluck_num_parameters}{tidy_xrf}(x, ...)

\method{.pluck_num_parameters}{xrf}(x, penalty = 0.001, ...)

\method{.pluck_num_parameters}{bagger}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{committees}{The number of Cubist committees to use in the examination.}

\item{penalty}{A numeric parameter value for the amount of
regularization/penalty of the model.}
}
\description{
Compute the number of estimated model parameters
}
\keyword{internal}
