% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_rules.R
\name{pluck_num_rules}
\alias{pluck_num_rules}
\alias{.pluck_num_rules}
\alias{.pluck_num_rules.default}
\alias{.pluck_num_rules.workflow}
\alias{.pluck_num_rules.model_fit}
\alias{.pluck_num_rules.tidy_C50}
\alias{.pluck_num_rules.C5.0}
\alias{.pluck_num_rules.tidy_cubist}
\alias{.pluck_num_rules.cubist}
\alias{.pluck_num_rules.xrf}
\title{Compute the number of rules in a rule-based model}
\usage{
.pluck_num_rules(x, ...)

\method{.pluck_num_rules}{default}(x, ...)

\method{.pluck_num_rules}{workflow}(x, ...)

\method{.pluck_num_rules}{model_fit}(x, ...)

\method{.pluck_num_rules}{tidy_C50}(x, trials = max(x$trial), ...)

\method{.pluck_num_rules}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{.pluck_num_rules}{tidy_cubist}(x, committees = max(x$committee), ...)

\method{.pluck_num_rules}{cubist}(x, committees = x$committees, ...)

\method{.pluck_num_rules}{xrf}(x, penalty = 0.001, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{committees}{The number of Cubist committees to use in the examination.}

\item{penalty}{A numeric parameter value for the amount of
regularization/penalty of the model.}
}
\description{
Compute the number of rules in a rule-based model
}
\keyword{internal}
