% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_support_vectors.R
\name{pluck_num_support_vectors}
\alias{pluck_num_support_vectors}
\alias{.pluck_num_support_vectors}
\alias{.pluck_num_support_vectors.default}
\alias{.pluck_num_support_vectors.workflow}
\alias{.pluck_num_support_vectors.model_fit}
\alias{.pluck_num_support_vectors.ksvm}
\title{Number of total support vectors used by kernel model}
\usage{
.pluck_num_support_vectors(x, ...)

\method{.pluck_num_support_vectors}{default}(x, ...)

\method{.pluck_num_support_vectors}{workflow}(x, ...)

\method{.pluck_num_support_vectors}{model_fit}(x, ...)

\method{.pluck_num_support_vectors}{ksvm}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}
}
\description{
Number of total support vectors used by kernel model
}
\keyword{internal}
