% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_term_nodes.R
\name{pluck_num_term_nodes}
\alias{pluck_num_term_nodes}
\alias{.pluck_num_term_nodes}
\alias{.pluck_num_term_nodes.default}
\alias{.pluck_num_term_nodes.workflow}
\alias{.pluck_num_term_nodes.model_fit}
\alias{.pluck_num_term_nodes.rpart}
\alias{.pluck_num_term_nodes.tidy_C50}
\alias{.pluck_num_term_nodes.C5.0}
\alias{.pluck_num_term_nodes.ranger}
\alias{.pluck_num_term_nodes.xgb.Booster}
\alias{.pluck_num_term_nodes.partynode}
\alias{.pluck_num_term_nodes.party}
\alias{.pluck_num_term_nodes.cforest}
\alias{.pluck_num_term_nodes.bagger}
\alias{.pluck_num_term_nodes.bart}
\alias{.pluck_num_term_nodes.lgb_trees}
\alias{.pluck_num_term_nodes.lgb.Booster}
\title{Compute the number of terminal nodes in a tree-based model}
\usage{
.pluck_num_term_nodes(x, ...)

\method{.pluck_num_term_nodes}{default}(x, ...)

\method{.pluck_num_term_nodes}{workflow}(x, ...)

\method{.pluck_num_term_nodes}{model_fit}(x, ...)

\method{.pluck_num_term_nodes}{rpart}(x, ...)

\method{.pluck_num_term_nodes}{tidy_C50}(x, trials = max(x$trial), ...)

\method{.pluck_num_term_nodes}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{.pluck_num_term_nodes}{ranger}(x, ...)

\method{.pluck_num_term_nodes}{xgb.Booster}(x, nrounds = x$niter, ...)

\method{.pluck_num_term_nodes}{partynode}(x, ...)

\method{.pluck_num_term_nodes}{party}(x, ...)

\method{.pluck_num_term_nodes}{cforest}(x, ...)

\method{.pluck_num_term_nodes}{bagger}(x, ...)

\method{.pluck_num_term_nodes}{bart}(x, ...)

\method{.pluck_num_term_nodes}{lgb_trees}(x, trees = max(x$trees), ...)

\method{.pluck_num_term_nodes}{lgb.Booster}(x, trees = x$params$num_iterations, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{trials, nrounds, trees}{The number of boosting to use in the characterization.}
}
\description{
Compute the number of terminal nodes in a tree-based model
}
\keyword{internal}
