% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_tune.R
\name{retain_characteristics}
\alias{retain_characteristics}
\alias{collect_characteristics}
\title{tune helpers}
\usage{
retain_characteristics(wflow)

collect_characteristics(x, summarize = TRUE, add_metrics = FALSE, wide = FALSE)
}
\arguments{
\item{wflow}{A fitted model workflow.}

\item{x}{The results of \code{\link[tune:tune_grid]{tune_grid()}}, \code{\link[tune:tune_bayes]{tune_bayes()}}, \code{\link[tune:fit_resamples]{fit_resamples()}},
or \code{\link[tune:last_fit]{last_fit()}}. For \code{\link[tune:collect_predictions]{collect_predictions()}}, the control option \code{save_pred = TRUE} should have been used.}

\item{summarize}{A logical; should metrics be summarized over resamples
(\code{TRUE}) or return the values for each individual resample. Note that, if \code{x}
is created by \code{\link[tune:last_fit]{last_fit()}}, \code{summarize} has no effect. For the other object
types, the method of summarizing predictions is detailed below.}

\item{add_metrics}{A logical to combine the results the associated
performance metrics.}

\item{wide}{A logical for making some information in wide format.}
}
\value{
A tibble.
}
\description{
tune helpers
}
\details{
When using the \pkg{tune} or \pkg{finetune} packages,
\code{retain_characteristics()} can be used with the resampling and tuning functions
to return the characteristics from multiple model configurations. These
results are contained in a column called \code{.extracts}.

\code{collect_characteristics()} can be used to easily wrangle the results in the
\code{.extracts} column into a tibble of usable results.

The vignette \emph{Using characterize with tidymodels} shows an example.
}
