% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_features.R
\name{pluck_active_features}
\alias{pluck_active_features}
\alias{.pluck_active_features}
\alias{.pluck_active_features.default}
\alias{.pluck_active_features.workflow}
\alias{.pluck_active_features.model_fit}
\alias{.pluck_active_features.terms}
\alias{.pluck_active_features.rpart}
\alias{.pluck_active_features.ranger}
\alias{.pluck_active_features.xgb.Booster}
\alias{.pluck_active_features.glmnet}
\alias{.pluck_active_features.cubist}
\alias{.pluck_active_features.C5.0}
\alias{.pluck_active_features.earth}
\alias{.pluck_active_features.mixo_spls}
\alias{.pluck_active_features.mixo_pls}
\alias{.pluck_active_features.mixo_splsda}
\alias{.pluck_active_features.mixo_plsda}
\alias{.pluck_active_features.bagger}
\alias{.pluck_active_features.bart}
\alias{.pluck_active_features.party}
\alias{.pluck_active_features.cforest}
\alias{.pluck_active_features.xrf}
\title{Predictors used in the model}
\usage{
.pluck_active_features(x, ...)

\method{.pluck_active_features}{default}(x, ...)

\method{.pluck_active_features}{workflow}(x, ...)

\method{.pluck_active_features}{model_fit}(x, ...)

\method{.pluck_active_features}{terms}(x, ...)

\method{.pluck_active_features}{rpart}(x, ...)

\method{.pluck_active_features}{ranger}(x, ...)

\method{.pluck_active_features}{xgb.Booster}(x, nrounds = x$niter, ...)

\method{.pluck_active_features}{glmnet}(x, penalty = 0.001, ...)

\method{.pluck_active_features}{cubist}(x, committees = x$committees, ...)

\method{.pluck_active_features}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{.pluck_active_features}{earth}(x, ...)

\method{.pluck_active_features}{mixo_spls}(x, ...)

\method{.pluck_active_features}{mixo_pls}(x, ...)

\method{.pluck_active_features}{mixo_splsda}(x, ...)

\method{.pluck_active_features}{mixo_plsda}(x, ...)

\method{.pluck_active_features}{bagger}(x, ...)

\method{.pluck_active_features}{bart}(x, ...)

\method{.pluck_active_features}{party}(x, ...)

\method{.pluck_active_features}{cforest}(x, ...)

\method{.pluck_active_features}{xrf}(x, penalty = 0.001, ...)

\method{.pluck_active_features}{bagger}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{penalty}{A numeric parameter value for the amount of
regularization/penalty of the model.}

\item{committees}{The number of Cubist committees to use in the examination.}

\item{trials, nrounds}{The number of boosting to use in the examination.}
}
\description{
Predictors used in the model
}
\keyword{internal}
