% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_active_features.R
\name{pluck_num_active_features}
\alias{pluck_num_active_features}
\alias{.pluck_num_active_features}
\alias{.pluck_num_active_features.default}
\alias{.pluck_num_active_features.workflow}
\alias{.pluck_num_active_features.model_fit}
\alias{.pluck_num_active_features.glmnet}
\alias{.pluck_num_active_features.cubist}
\alias{.pluck_num_active_features.C5.0}
\alias{.pluck_num_active_features.xrf}
\title{Number of predictors used in the model}
\usage{
.pluck_num_active_features(x, ...)

\method{.pluck_num_active_features}{default}(x, ...)

\method{.pluck_num_active_features}{workflow}(x, ...)

\method{.pluck_num_active_features}{model_fit}(x, ...)

\method{.pluck_num_active_features}{glmnet}(x, penalty = 0.001, ...)

\method{.pluck_num_active_features}{cubist}(x, committees = x$committees, ...)

\method{.pluck_num_active_features}{C5.0}(x, trials = x$trials["Actual"], ...)

\method{.pluck_num_active_features}{xrf}(x, penalty = 0.001, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Not currently used.}

\item{penalty}{A numeric parameter value for the amount of
regularization/penalty of the model.}

\item{committees}{The number of Cubist committees to use in the examination.}
}
\description{
Number of predictors used in the model
}
\keyword{internal}
